# openapi.yaml
openapi: 3.0.3
info:
  title: Portal de Benefícios KNN API
  description: |
    API para o Portal de Benefícios KNN, permitindo que alunos da KNN Idiomas acessem promoções
    exclusivas de parceiros comerciais através de códigos de validação temporários.
  version: 1.0.0
  contact:
    name: Equipe de TI KNN Idiomas
    email: ti@knnidiomas.com.br
servers:
  - url: https://api.knnidiomas.com.br/v1
    description: Ambiente de produção
  - url: http://localhost:8000/v1
    description: Ambiente de desenvolvimento local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT com claim 'role' (student|partner|admin)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_CODE
            msg:
              type: string
              example: Código expirado ou já usado
          required:
            - code
            - msg
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        data:
          type: object
          description: Dados da resposta (varia conforme endpoint)
        msg:
          type: string
          default: ok
      required:
        - data
        - msg
...
    Partner:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trade_name:
          type: string
          example: Livraria Cultura
        category:
          type: string
          example: Livraria
        address:
          type: string
          example: Av. Paulista, 2073 - São Paulo/SP
        active:
          type: boolean
          default: true
      required:
        - id
        - trade_name
        - category
        - address
        - active
    PartnerDetail:
      allOf:
        - $ref: '#/components/schemas/Partner'
        - type: object
          properties:
            promotions:
              type: array
              items:
                $ref: '#/components/schemas/Promotion'
    Promotion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        partner_id:
          type: string
          format: uuid
        title:
          type: string
          example: 20% de desconto em livros de idiomas
        type:
          type: string
          example: discount
        valid_from:
          type: string
          format: date-time
        valid_to:
          type: string
          format: date-time
        active:
          type: boolean
          default: true
      required:
        - id
        - partner_id
        - title
        - type
        - valid_from
        - valid_to
        - active
...
    ValidationCodeRequest:
      type: object
      properties:
        partner_id:
          type: string
          format: uuid
      required:
        - partner_id
...
    RedemptionHistory:
      type: array
      items:
        type: object
        properties:
          id:
            type: string
            format: uuid
          partner:
            type: object
            properties:
              id:
                type: string
                format: uuid
              trade_name:
                type: string
            required:
              - id
              - trade_name
          promotion:
            type: object
            properties:
              title:
                type: string
            required:
              - title
          value:
            type: number
            format: float
          used_at:
            type: string
            format: date-time
        required:
          - id
          - partner
          - promotion
          - value
          - used_at
    FavoritePartners:
      type: array
      items:
        $ref: '#/components/schemas/Partner'
  parameters:
    CategoryParam:
      name: cat
      in: query
      description: Filtro por categoria de parceiro
      schema:
        type: string
    OrderParam:
      name: ord
      in: query
      description: Ordenação dos resultados
      schema:
        type: string
        enum: [name_asc, name_desc, category_asc, category_desc]
    PaginationLimit:
      name: limit
      in: query
      description: Número máximo de itens por página
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    PaginationOffset:
      name: offset
      in: query
      description: Offset para paginação
      schema:
        type: integer
        default: 0
        minimum: 0
...
    IdParam:
      name: id
      in: path
      description: ID do recurso
      required: true
      schema:
        type: string
        format: uuid
    PartnerIdParam:
      name: pid
      in: path
      description: ID do parceiro
      required: true
      schema:
        type: string
        format: uuid
security:
  - bearerAuth: []
paths:
...
  /partners:
    get:
      summary: Lista parceiros
      description: Retorna lista de parceiros com filtros e paginação
      tags:
        - student
      parameters:
        - $ref: '#/components/parameters/CategoryParam'
        - $ref: '#/components/parameters/OrderParam'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Lista de parceiros
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Partner'
                          total:
                            type: integer
                          limit:
                            type: integer
                          offset:
                            type: integer
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /partners/{id}:
    get:
      summary: Detalhes do parceiro
      description: Retorna detalhes do parceiro e suas promoções ativas
      tags:
        - student
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Detalhes do parceiro
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PartnerDetail'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Parceiro não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /validation-codes:
    post:
      summary: Gera código de validação
      description: Gera um código de validação de 6 dígitos que expira em 3 minutos
      tags:
        - student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationCodeRequest'
      responses:
        '201':
          description: Código gerado com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          code:
                            type: string
                            example: '123456'
                          expires:
                            type: string
                            format: date-time
                        required:
                          - code
                          - expires
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Parceiro não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /students/me/history:
    get:
      summary: Histórico de resgates
      description: Retorna o histórico de resgates do aluno
      tags:
        - student
      parameters:
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Histórico de resgates
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            $ref: '#/components/schemas/RedemptionHistory'
                          total:
                            type: integer
                          limit:
                            type: integer
                          offset:
                            type: integer
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /students/me/fav:
    get:
      summary: Lista parceiros favoritos
      description: Retorna a lista de parceiros favoritos do aluno
      tags:
        - student
      responses:
        '200':
          description: Lista de parceiros favoritos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FavoritePartners'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Adiciona parceiro aos favoritos
      description: Adiciona um parceiro à lista de favoritos do aluno
      tags:
        - student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                partner_id:
                  type: string
                  format: uuid
              required:
                - partner_id
      responses:
        '201':
          description: Parceiro adicionado aos favoritos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Parceiro não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /students/me/fav/{pid}:
    delete:
      summary: Remove parceiro dos favoritos
      description: Remove um parceiro da lista de favoritos do aluno
      tags:
        - student
      parameters:
        - $ref: '#/components/parameters/PartnerIdParam'
      responses:
        '200':
          description: Parceiro removido dos favoritos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Parceiro não encontrado nos favoritos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...